changed to fixed threads, fixed level

+++ CCAS.h
@@ -13,10 +13,10 @@ enum STATUS
 class CCAS{
 private:
     typedef struct CCASDESCRIPTOR{
-        WORD *a;
-        WORD e, n;
+        int64 *a;
+        int64 e, n;
         STATUS *cond;
-        CCASDESCRIPTOR(WORD *_a, WORD _e, WORD _n, STATUS *_cond) 
+        CCASDESCRIPTOR(int64 *_a, int64 _e, int64 _n, STATUS *_cond) 
                 : a(_a), e(_e), n(_n), cond(_cond) {}
     }CCASDesc;
 
@@ -24,9 +24,9 @@ public:
     CCAS();
     ~CCAS();
 
-    bool IsCCASDesc(WORD d);
-    void doCCAS(WORD *a, WORD e, WORD n, STATUS *cond);
-    WORD CCASRead(WORD *a);
+    bool IsCCASDesc(int64 d);
+    void doCCAS(int64 *a, int64 e, int64 n, STATUS *cond);
+    int64 CCASRead(int64 *a);
     void CCASHelp(CCASDesc *d);
 };
 
@@ -36,31 +36,31 @@ CCAS::CCAS(){
 CCAS::~CCAS(){
 }
 
-void CCAS::doCCAS(WORD *a, WORD e, WORD n, STATUS *cond){
+void CCAS::doCCAS(int64 *a, int64 e, int64 n, STATUS *cond){
     CCASDesc *d = new CCASDesc(a,e,n,cond);
-    bool v = __sync_bool_compare_and_swap(d->a,d->e,((WORD)d)|2);
+    bool v = __sync_bool_compare_and_swap(d->a,d->e,((int64)d)|2);
     while(!v){
         if(!IsCCASDesc(v)) return;
         CCASHelp((CCASDesc *)v);
-        v = __sync_val_compare_and_swap(d->a,d->e,((WORD)d)|2);
+        v = __sync_val_compare_and_swap(d->a,d->e,((int64)d)|2);
     }
     CCASHelp(d);
 }
 
-WORD CCAS::CCASRead (WORD *a){
-    WORD v;
+int64 CCAS::CCASRead (int64 *a){
+    int64 v;
     for(v = *a; IsCCASDesc(v); v = *a)
         CCASHelp((CCASDesc *)v);
     return v;
 }
 
 void CCAS::CCASHelp(CCASDesc *d){
-    CCASDesc *dd = (CCASDesc *)((WORD)d & (~2));
-    WORD dx = (WORD)d|2;
+    CCASDesc *dd = (CCASDesc *)((int64)d & (~2));
+    int64 dx = (int64)d|2;
     bool success = *(dd->cond) == UNDECIDED;
     bool v = __sync_bool_compare_and_swap(dd->a,dx,success?dd->n:dd->e);
 }
 
-bool CCAS::IsCCASDesc(WORD d){
-    return (WORD)d & 2;
+bool CCAS::IsCCASDesc(int64 d){
+    return (int64)d & 2;
 }
